renamed encoders with Spark prefix and added to spark package
