+++ simulation/SyntheSimJava/src/main/java/com/autodesk/synthesis/revrobotics/spark/SparkAbsoluteEncoder.java
